# api/openapi.yaml
openapi: 3.1.0
info:
  title: NXT TIME PULSE Admin API
  version: 1.0.0
  description: Admin-gated API endpoints for GEO TIMESHEET - TIMEPULSE
servers:
  - url: /api
    description: API base path
security:
  - bearerAuth: []
paths:
...
  /users/{id}:
    patch:
      summary: Update user role/status (admin/manager only)
      tags: [Users]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserUpdate" }
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400":
          description: Validation failed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: User not found
    delete:
      summary: Remove user from organization (admin/manager only)
      tags: [Users]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: User removed
        "400":
          description: Cannot remove yourself
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: User not found
...
  /users/{id}/offboard:
    post:
      summary: Request user offboarding (admin only, maker-checker)
      tags: [Users]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string }
                effectiveDate: { type: string, format: date }
      responses:
        "202":
          description: Offboarding request queued for approval
        "400":
          description: Validation or self-offboarding blocked
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: User not found
...
  /rates:
    get:
      summary: List rate cards (admin-only compensation visibility)
      tags: [Rates]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: userId
          in: query
          schema: { type: string }
        - name: projectId
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string, enum: [pending, active, superseded, cancelled] }
      responses:
        "200":
          description: Rate cards
    post:
      summary: Submit rate-card create request (maker-checker)
      tags: [Rates]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, payRate, currency, effectiveFrom]
              properties:
                userId: { type: string }
                projectId: { type: string, nullable: true }
                payRate: { type: number }
                billRate: { type: number, nullable: true }
                currency: { type: string, minLength: 3, maxLength: 3 }
                effectiveFrom: { type: string, format: date }
                effectiveTo: { type: string, format: date, nullable: true }
                changeReason: { type: string, nullable: true }
      responses:
        "202":
          description: Rate change queued for approval
...
  /reports/payroll:
    get:
      summary: Export payroll CSV (admin/manager only)
      tags: [Reports]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: from
          in: query
          required: true
          schema: { type: string, format: date }
        - name: to
          in: query
          required: true
          schema: { type: string, format: date }
      responses:
        "200":
          description: CSV file
          content:
            text/csv: {}
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string, nullable: true }
        firstName: { type: string, nullable: true }
        lastName: { type: string, nullable: true }
        role: { type: string }
        status: { type: string }
        _count:
          type: object
          properties:
            timesheets: { type: integer }
            allocations: { type: integer }
    UserUpdate:
      type: object
      properties:
        role: { type: string, enum: [admin, manager, worker] }
        status: { type: string, enum: [invited, active, suspended, offboarded, archived] }
      minProperties: 1
    Error:
      type: object
      properties:
        error: { type: string }
        details: { type: object }
  responses:
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string, example: "Unauthorized" }
    Forbidden:
      description: Insufficient permissions (requires org:admin or org:manager)
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string, example: "Forbidden" }